moved message handlers from background/index.ts to messaging.ts
